+++ app/views/__init__.py
@@ -1,0 +1,2 @@
+'''import Blueprints to views module'''
+from flask import Blueprint
