+++ book_a_meal.py
@@ -15,5 +15,6 @@ APP = create_app(CONFIG)
 
 
 if __name__ == '__main__':
-    PORT = int(os.environ.get('PORT', 5000))
-    APP.run('', port=PORT)
+    # PORT = int(os.environ.get('PORT', 5000))
+    # APP.run('', port=PORT, debug=True)
+    APP.run()
